refactor: Make snapshot module more self-contained
Move Rollback component and SaveWorld, LoadWorld, AdvanceWorld scheduel
labels into snapshot plugin.
