+++ src/snapshot/rollback.rs
